docs(examples): add testable examples for Hide and SetHidden
